Add embedWindowTestUtils and refactor related test files
